+++ src/main/java/info/kfgodel/bean2bean/converters/optionals/Optional2ObjectConverter.java
@@ -1,4 +1,4 @@
-package info.kfgodel.bean2bean.converters;
+package info.kfgodel.bean2bean.converters.optionals;
 
 import info.kfgodel.bean2bean.core.api.Bean2beanTask;
