+++ src/test/java/info/kfgodel/bean2bean/converters/Array2ArrayConverterTest.java
@@ -2,6 +2,8 @@ package info.kfgodel.bean2bean.converters;
 
 import ar.com.dgarcia.javaspec.api.JavaSpec;
 import ar.com.dgarcia.javaspec.api.JavaSpecRunner;
+import info.kfgodel.bean2bean.converters.collections.Array2ArrayConverter;
+import info.kfgodel.bean2bean.converters.generators.ArrayInstantiator;
 import info.kfgodel.bean2bean.core.api.exceptions.ConversionException;
 import info.kfgodel.bean2bean.core.api.exceptions.NestedConversionException;
 import info.kfgodel.bean2bean.dsl.impl.Dsl;
