+++ src/test/java/info/kfgodel/bean2bean/converters/EnumConvertersTest.java
@@ -2,6 +2,8 @@ package info.kfgodel.bean2bean.converters;
 
 import ar.com.dgarcia.javaspec.api.JavaSpec;
 import ar.com.dgarcia.javaspec.api.JavaSpecRunner;
+import info.kfgodel.bean2bean.converters.enums.Enum2StringConverter;
+import info.kfgodel.bean2bean.converters.enums.String2EnumConverter;
 import info.kfgodel.bean2bean.core.api.exceptions.ConversionException;
 import info.kfgodel.bean2bean.dsl.impl.Dsl;
 import org.junit.runner.RunWith;
